feat: align core schema constraints
- 도서 ISBN과 로케이션 바코드를 unique 제약으로 명시했습니다.
- books.publisher를 추가했습니다.
- 가격 관련 필드는 문서에 따라 int에서 DECIMAL로 변경하였으며, 해당 필드는 다음과 같습니다
: books.base_price, orders.total_price, order_items.unit_price, order_items.final_price

- user의 employee_id, email을 필수 unique 값으로 명시했습니다.
- must_change_password 필드를 추가했습니다.
(팀원 요청에 따라, 초기 비밀번호 변경 플로우와 사번 기반 사용자 식별을 위한 기반 작업.)

위의 변경에 따라 SQLModel 모델과 db/init.sql이 서로 다른 스키마를 바라보지 않도록 DDL도 함께 갱신했으며,
unique 제약 추가로 인해 mock seed가 반복 실행 시 충돌하지 않도록 seed 데이터를 유니크하게 생성하도록 조정.

+++ app/api/routes/orders.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from typing import List
 from uuid import UUID
 
@@ -23,7 +24,7 @@ class CreateOrderRequest(BaseModel):
 
 class CreateOrderResponse(BaseModel):
     order_id: UUID
-    total_price: int
+    total_price: Decimal
     applied_discount: str  # 할인 적용되었는지 상태를 나타내는 str 값. 
     status: OrderStatus
